feat(build): Add ARM64 platform support
The build system now includes `linux/arm64` in the common set of
platforms for most targets. This expands the architecture support
for the generated Docker images.

However, the `extra-targets` group explicitly overrides this to
exclude `linux/arm64` due to known build issues with certain
components, such as Citus, on the ARM64 architecture.

Signed-off-by: Enrique Hernández Bello <enrique.hernandez@maarlab.com>